meta-integrity: import gpg key before do_package_write_rpm
os-release is almost the first target recipe built for a rootfs.
However, there is no dependency making sure gpg key is imported before
signing os-release recipe.

This commit defines that dependency and fixes the problem above.

Also make the signing-keys into a native recipe, as it is meanless to be
a target recipe.

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>