meta-integrity, meta-signing-key: Make the secondary key populate
Today we have an "extra_system_trusted_key" that is created by the
create-user-key-store script.  We don't however provide this as a DER
encoded certificate for use with the kernel keyrings.  Provide this now
in key-store-extra-system-trusted-cert.  In the IMA init script, if we
have a secondary keyring available in the kernel and we have any
secondary certificate in a DER encoding, load them into the secondary
keyring.

Signed-off-by: Tom Rini <trini@konsulko.com>